Enhance Movie CRUD Operations with Service Repository (#4)
* Move context file

* Add Generic Repository Interfaces and Implementation

The `IEntityBase` interface represents a common structure for entities with an ID. The `IEntityBaseRepository` interface defines standard CRUD operations for any entity that implements `IEntityBase`. The `EntityBaseRepository` class is the concrete implementation of the generic repository interface providing basic CRUD async methods using EntityFramework.

* Introduce movie view model

Note: The `GenreIds` property is added to store the selected genre IDs from the view form.

* Refactor Movie Class to Implement IEntityBase

* Introduce Movies Service

The interface includes a new method, AddMovieVMAsync, to asynchronously add a movie based on the provided MovieVM.

* Update Movie-Genre Relationship Configuration

This commit includes changes to the relationship configuration within the movie_has_genre table, ensuring proper cascade behaviors for both movies and genres.

* Introduce Movie Services

Add IMoviesService and MoviesService with methods for asynchronous movie addition and updating (`AddMovieAsync`, `UpdateMovieAsync`). Override `GetByIdAsync` and `GetAllAsync` in MoviesService. Configure service registration in Program.cs.

* Change Update and Delete Operations

* Add Genre Service and Interface

Introduce `IGenresService` interface and its implementation `GenresService` to manage genre-related operations.

* Register Genre Service in Dependency Injection

* Add Movie Mapper

Add MovieMapper class responsible for mapping between Movie entities and MovieVM view models.

* Refactor Movies Controller

Refactor MoviesController to improve code organization and readability. Utilize `IMoviesService` for movie operations, `IGenresService` for genre retrieval, and `MovieMapper` for entity-to-view-model conversions.

* Display movie genres to index

* Update Edit View for Movie

Modify the Edit.cshtml view to use MovieVM as the model type. Change the ReleaseDate input to a date type and introduce a multi-select dropdown for GenreIds.

* Update Details View for Movie

Display the associated genres separated by commas.

* Update Create View for Movie

Modify the Create.cshtml view to use MovieVM as the model type. Change the ReleaseDate input to a date type and introduce a multi-select dropdown for GenreIds.

* Reorganize Folder Structure

File relocation, directory creation, and updates to namespaces and references are part of this change.